rename file for client_entity_map
